+++ src/ldc/filter/_randomize_records.py
@@ -1,0 +1,128 @@
+import argparse
+from typing import List
+from random import Random, shuffle
+
+from wai.logging import LOGGING_WARNING
+from ldc.core import DOMAIN_PAIRS, DOMAIN_PRETRAIN, DOMAIN_TRANSLATION
+from ._core import Filter
+from ldc.pretrain import PretrainData
+from ldc.supervised.pairs import PairData
+from ldc.translation import TranslationData
+
+
+class RandomizeRecords(Filter):
+    """
+    Batch filter that randomizes the order of the records.
+    """
+
+    def __init__(self, seed: int = 1, logger_name: str = None, logging_level: str = LOGGING_WARNING):
+        """
+        Initializes the filter.
+
+        :param seed: the seed value to use
+        :type seed: int
+        :param logger_name: the name to use for the logger
+        :type logger_name: str
+        :param logging_level: the logging level to use
+        :type logging_level: str
+        """
+        super().__init__(logger_name=logger_name, logging_level=logging_level)
+        self.seed = seed
+        self._rnd = None
+
+    def name(self) -> str:
+        """
+        Returns the name of the handler, used as sub-command.
+
+        :return: the name
+        :rtype: str
+        """
+        return "randomize-records"
+
+    def description(self) -> str:
+        """
+        Returns a description of the reader.
+
+        :return: the description
+        :rtype: str
+        """
+        return "Batch filter that randomizes the order of the records."
+
+    def domains(self) -> List[str]:
+        """
+        Returns the domains of the filter.
+
+        :return: the domains
+        :rtype: list
+        """
+        return [DOMAIN_PAIRS, DOMAIN_PRETRAIN, DOMAIN_TRANSLATION]
+
+    def accepts(self) -> List:
+        """
+        Returns the list of classes that are accepted.
+
+        :return: the list of classes
+        :rtype: list
+        """
+        return [PairData, PretrainData, TranslationData]
+
+    def generates(self) -> List:
+        """
+        Returns the list of classes that get produced.
+
+        :return: the list of classes
+        :rtype: list
+        """
+        return [PairData, PretrainData, TranslationData]
+
+    def _create_argparser(self) -> argparse.ArgumentParser:
+        """
+        Creates an argument parser. Derived classes need to fill in the options.
+
+        :return: the parser
+        :rtype: argparse.ArgumentParser
+        """
+        parser = super()._create_argparser()
+        parser.add_argument("-s", "--seed", type=int, help="The seed value to use for the randomization. Without seed value the order will differ between runs.", default=None, required=False)
+        return parser
+
+    def _apply_args(self, ns: argparse.Namespace):
+        """
+        Initializes the object with the arguments of the parsed namespace.
+
+        :param ns: the parsed arguments
+        :type ns: argparse.Namespace
+        """
+        super()._apply_args(ns)
+        self.seed = ns.seed
+
+    def initialize(self):
+        """
+        Initializes the processing, e.g., for opening files or databases.
+        """
+        super().initialize()
+        self._rnd = Random(self.seed)
+
+    def _requires_list_input(self) -> bool:
+        """
+        Returns whether lists are expected as input for the _process method.
+
+        :return: True if list inputs are expected by the filter
+        :rtype: bool
+        """
+        return True
+
+    def _do_process(self, data):
+        """
+        Processes the data record(s).
+
+        :param data: the record(s) to process
+        :return: the potentially updated record or None if to drop
+        """
+        if isinstance(data, list):
+            result = list(data)
+            self._rnd.shuffle(result)
+            return result
+        else:
+            self.logger().warning("Can only randomize lists, but received: %s" % str(type(data)))
+            return data