+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 0.0.5 (????-??-??)
 ------------------
 
-- ...
+- added flag `-b/--force_batch` to the `llm-convert` tool which all data to be reader from the
+  reader before filtering it and then passing it to the writer; useful for batch filters.
+- added the `randomize-records` batch filter
 
 
 0.0.4 (2023-12-19)